AI Model/requirements.txt
torch
transformers
accelerate
mysql-connector-python
beautifulsoup4
requests

pip install -r requirements.txt
